requirements.txt
pyserial
PyGObject
dbus-python
numpy
pyusb
gatt
supervisor
luma.oled